+++ packages/vibrant-components/src/lib/TableFooter/translation.ts
@@ -1,0 +1,16 @@
+import type { Translation } from '@vibrant-ui/core';
+
+export const ko: Translation['tableFooter'] = {
+  rowsPerPage: '페이지당 행 수',
+  total: '총 {count}개',
+};
+
+export const en: Translation['tableFooter'] = {
+  rowsPerPage: 'Rows per page',
+  total: 'total {count} items',
+};
+
+export const ja: Translation['tableFooter'] = {
+  rowsPerPage: '1 ページあたりの行数',
+  total: '合計 {count} 個',
+};