+++ packages/vibrant-core/src/lib/ScrollBox/ScrollBox.tsx
@@ -1,10 +1,11 @@
 import { createElement, forwardRef } from 'react';
 import styled from '@emotion/styled';
+import { transformResponsiveValue } from '../transformResponsiveValue';
 import type { ScrollBoxProps } from './ScrollBoxProps';
 import { interpolation, shouldForwardProp } from './ScrollBoxProps';
 
-export const ScrollBox = styled(
-  forwardRef<any, ScrollBoxProps>(({ as = 'div', keyboardShouldPersistTaps: _, ...restProps }, ref) =>
+const SystemScrollBox = styled(
+  forwardRef<any, ScrollBoxProps>(({ as = 'div', ...restProps }, ref) =>
     createElement(as, {
       ref,
       ...restProps,
@@ -14,5 +15,15 @@ export const ScrollBox = styled(
     shouldForwardProp,
   }
 )(interpolation);
+
+export const ScrollBox = forwardRef<any, ScrollBoxProps>(
+  ({ keyboardShouldPersistTaps: _, scrollEnabled = true, ...restProps }, ref) => (
+    <SystemScrollBox
+      ref={ref}
+      {...restProps}
+      overflow={transformResponsiveValue(scrollEnabled, value => (value ? 'auto' : 'hidden'))}
+    />
+  )
+);
 
 ScrollBox.displayName = 'ScrollBox';